ajout du delete book usecase dans l api

+++ libs/good-place/src/application/usecases/delete-book.usecase.ts
@@ -1,10 +1,16 @@
+import { Injectable } from '@nestjs/common';
 import { BookRepository } from '../book.repository';
 
+@Injectable()
 export class DeleteBookUseCase {
   constructor(private readonly bookRepository: BookRepository) {}
 
   async handle(id: string) {
     //TODO: ajouter la logique de suppression si owner.
+    const existingBook = await this.bookRepository.getBookById(id);
+    if (!existingBook) {
+      return Promise.resolve();
+    }
     await this.bookRepository.deleteBookById(id);
   }
 }
